requirements.txt
matplotlib
torch 
torchvision
numpy
opencv-python
Pillow
beautifulsoup4
requests